+++ src/Client.php
@@ -3,6 +3,7 @@
 namespace Darmen\AzureFace;
 
 use Darmen\AzureFace\Http\Middleware;
+use Darmen\AzureFace\Resources\Face;
 use Darmen\AzureFace\Resources\FaceList;
 use Darmen\AzureFace\Resources\LargeFaceList;
 use GuzzleHttp\Client as HttpClient;
@@ -11,6 +12,7 @@ use GuzzleHttp\HandlerStack;
 /**
  * @method LargeFaceList largeFaceList()
  * @method FaceList faceList()
+ * @method Face face()
  */
 class Client
 {