+++ tests/ClientTest.php
@@ -3,6 +3,7 @@
 namespace Darmen\AzureFace\Tests;
 
 use Darmen\AzureFace\Client;
+use Darmen\AzureFace\Resources\Face;
 use Darmen\AzureFace\Resources\FaceList;
 use Darmen\AzureFace\Resources\LargeFaceList;
 use PHPUnit\Framework\TestCase;
@@ -27,6 +28,14 @@ class ClientTest extends TestCase
         $this->assertEquals(
             get_class($this->sut->largeFaceList()),
             LargeFaceList::class
+        );
+    }
+
+    public function testFaceReturnsFace(): void
+    {
+        $this->assertEquals(
+            get_class($this->sut->face()),
+            Face::class
         );
     }
 }