+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+* Add `Face` resource
+
 ## [1.0.1] - 2021-07-27
 
 ### Fixed